rename module with measurements-related code
