+++ src/stats/common.rs
