+++ src/stats/io.rs
